feat: format code

+++ src/serialize/std_map_impl.hpp
@@ -18,7 +18,7 @@ template <typename T, typename std::enable_if<detail::is_map_like<T>::value, int
 serialize_iarchive& operator<<(T& t, serialize_iarchive& ia) {
   detail::size_type size;
   size << ia;
-  for (uint32_t i = 0; i < size.size; ++i) {
+  for (size_t i = 0; i < size.size; ++i) {
     typename T::value_type item;
     serialize_iarchive tmp;
     tmp << ia;
